Added Functionalities requested by TSE
