+++ src/variables.scss
@@ -31,6 +31,10 @@ $c-logo-gradient-1: #c0e6ff;
 $c-logo-gradient-2: #49d3ff;
 $c-logo-gradient-3: #a1e2ff;
 $c-logo: linear-gradient(263deg, $c-logo-gradient-1 37%, $c-logo-gradient-2 74%, $c-logo-gradient-3 100%);
+$c-line: rgba(255, 251, 0, .644);
+$c-important: #e5ff00;
+$c-very-important-1: #ff9d00;
+$c-very-important-2: #f35858;
 
 $space: 5px;
