+++ src/views/stylesheets/_taskFrm.scss
@@ -108,3 +108,7 @@ label {
     width: $space * 2;
   }
 }
+
+select {
+  margin: $space * 2;
+}
